+++ lib/core/network/api_services.dart
@@ -7,10 +7,11 @@ import 'package:flowers_ecommerce_app/features/auth/login/data/model/request/log
 import 'package:flowers_ecommerce_app/features/auth/login/data/model/responce/login_responce_dto.dart';
 import 'package:flowers_ecommerce_app/features/auth/register/data/model/register_body_dto.dart';
 import 'package:flowers_ecommerce_app/features/auth/register/data/model/register_respone/register_respone_dto.dart';
-import 'package:flowers_ecommerce_app/features/auth/reset_password/data/models/request/change_password_request_dto.dart';
-import 'package:flowers_ecommerce_app/features/auth/reset_password/data/models/response/change_password_response_dto.dart';
 import 'package:injectable/injectable.dart';
 import 'package:retrofit/retrofit.dart';
+
+import '../../features/auth/change_password/data/models/request/change_password_request_dto.dart';
+import '../../features/auth/change_password/data/models/response/change_password_response_dto.dart';
 
 part 'api_services.g.dart';
 
@@ -18,7 +19,7 @@ part 'api_services.g.dart';
 @injectable
 abstract class ApiServices {
   @factoryMethod
- factory ApiServices(Dio dio) = _ApiServices;
+  factory ApiServices(Dio dio) = _ApiServices;
   @POST(ApiConstants.loginEndpoint)
   Future<LoginResponceDto> login(@Body() LoginRequestDto loginRequestDto);
 
@@ -34,9 +35,8 @@ abstract class ApiServices {
   Future<RegisterResponeDto> register(@Body() RegisterBodyDTo request);
 
   @PATCH(ApiConstants.changePassword)
- Future<ChangePasswordResponseDto> changePassword(
-      @Body() ChangePasswordRequestDto requestDto,
-      @Header(ApiConstants.authorization) String token,
-      );
-
+  Future<ChangePasswordResponseDto> changePassword(
+    @Body() ChangePasswordRequestDto requestDto,
+    @Header(ApiConstants.authorization) String token,
+  );
 }
