+++ lib/features/auth/change_password/data/models/mapper/change_password_request_maper.dart
@@ -1,0 +1,9 @@
+import '../../../domain/entities/change_password_request_entity.dart';
+import '../request/change_password_request_dto.dart';
+
+ChangePasswordRequestDto toChangePasswordRequestDto(
+  ChangePasswordRequestEntity entity,
+) => ChangePasswordRequestDto(
+  password: entity.password,
+  newPassword: entity.newPassword,
+);
