+++ lib/features/auth/change_password/domain/entities/change_password_response_entity.dart
@@ -1,0 +1,11 @@
+import 'package:equatable/equatable.dart';
+
+class ChangePasswordResponseEntity extends Equatable {
+  const ChangePasswordResponseEntity({this.message, this.token});
+
+  final String? message;
+  final String? token;
+
+  @override
+  List<Object?> get props => [message, token];
+}
