+++ lib/features/auth/change_password/presentation/manager/change_password_event.dart
@@ -1,0 +1,3 @@
+sealed class ChangePasswordEvent {}
+
+class SubmitChangePasswordEvent extends ChangePasswordEvent {}
