--- lib/features/auth/reset_password/domain/entities/change_password_request_entity.dart
@@ -1,10 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class ChangePasswordRequestEntity extends Equatable{
-  final String? password;
-  final String? newPassword;
-
-  ChangePasswordRequestEntity({required this.password,required this.newPassword});
-
-  List<Object?> get props => [password,newPassword];
-}
