--- lib/features/auth/reset_password/domain/entities/change_password_response_entity.dart
@@ -1,14 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class ChangePasswordResponseEntity extends Equatable{
-  ChangePasswordResponseEntity({
-      this.message, 
-      this.token,});
-
-  String? message;
-  String? token;
-
-  @override
-  List<Object?> get props => [message,token];
-
-}
