--- lib/features/auth/reset_password/domain/repositories/change_password_repo.dart
@@ -1,7 +1,0 @@
-import 'package:flowers_ecommerce_app/core/errors/api_results.dart';
-import 'package:flowers_ecommerce_app/features/auth/reset_password/domain/entities/change_password_request_entity.dart';
-import 'package:flowers_ecommerce_app/features/auth/reset_password/domain/entities/change_password_response_entity.dart';
-
-abstract interface class ChangePasswordRepo{
-  Future<ApiResult<ChangePasswordResponseEntity>> changePassword(ChangePasswordRequestEntity request);
-}
