--- lib/features/auth/reset_password/presentation/manager/change_password_event.dart
@@ -1,3 +1,0 @@
-sealed class ChangePasswordEvent {}
-
-class SubmitChangePasswordEvent extends ChangePasswordEvent{}
